intelli-credit/backend/requirements.txt
django
django-ninja
requests
whitenoise
uvicorn
gunicorn
boto3
python-dotenv
python-jose
google-cloud-pubsub
